♻️ refactor(api): rename `interface` directory to `interfaces`
