+++ apps/api/src/application/use-cases/auth/index.ts
@@ -1,7 +1,7 @@
-export type { ILoginUseCase, ILoginUseCaseResult } from "./interface/login.usecase.interface";
-export type { IValidateSessionUseCase, IValidateSessionUseCaseResult } from "./interface/validate-session.usecase";
-export type { ISignupUseCase, ISignupUseCaseResult } from "./interface/signup.usecase.interface";
-export type { ILogoutUseCase } from "./interface/logout.usecase.interface";
+export type { ILoginUseCase, ILoginUseCaseResult } from "./interfaces/login.usecase.interface";
+export type { IValidateSessionUseCase, IValidateSessionUseCaseResult } from "./interfaces/validate-session.usecase";
+export type { ISignupUseCase, ISignupUseCaseResult } from "./interfaces/signup.usecase.interface";
+export type { ILogoutUseCase } from "./interfaces/logout.usecase.interface";
 export { LoginUseCase } from "./login.usecase";
 export { ValidateSessionUseCase } from "./validate-session.usecase";
 export { SignupUseCase } from "./signup.usecase";
