+++ apps/api/src/application/use-cases/auth/interfaces/signup.usecase.interface.ts
