+++ apps/api/src/application/use-cases/auth/interfaces/validate-session.usecase.ts
