+++ apps/api/src/application/use-cases/auth/logout.usecase.ts
@@ -1,6 +1,6 @@
 import type { ISessionRepository } from "../../../interface-adapter/repositories/session";
 import type { ISessionTokenService } from "../../../models/services/session-token";
-import type { ILogoutUseCase } from "./interface/logout.usecase.interface";
+import type { ILogoutUseCase } from "./interfaces/logout.usecase.interface";
 
 export class LogoutUseCase implements ILogoutUseCase {
 	constructor(
