+++ apps/api/src/application/use-cases/email-verification/interfaces/email-verification-confirm.usecase.interface.ts
