+++ apps/api/src/application/use-cases/email-verification/interfaces/email-verification-request.usecase.interface.ts
