+++ apps/api/src/application/use-cases/email/index.ts
@@ -1,3 +1,3 @@
 export { SendEmailUseCase } from "./send-email.usecase";
-export type { ISendEmailUseCase } from "./interface/send-email.usecase.interface";
+export type { ISendEmailUseCase } from "./interfaces/send-email.usecase.interface";
 export type { EmailRenderOptions, RequireAtLeastOne } from "./type";
