+++ apps/api/src/application/use-cases/email/interfaces/send-email.usecase.interface.ts
