+++ apps/api/src/application/use-cases/oauth/index.ts
@@ -1,12 +1,12 @@
 export type {
 	IOAuthLoginCallbackUseCase,
 	IOAuthLoginCallbackUseCaseResult,
-} from "./interface/oauth-login-callback.usecase.interface";
-export type { IOAuthRequestUseCase, IOAuthRequestUseCaseResult } from "./interface/oauth-request.usecase.interface";
+} from "./interfaces/oauth-login-callback.usecase.interface";
+export type { IOAuthRequestUseCase, IOAuthRequestUseCaseResult } from "./interfaces/oauth-request.usecase.interface";
 export type {
 	IOAuthSignupCallbackUseCase,
 	IOAuthSignupCallbackUseCaseResult,
-} from "./interface/oauth-signup-callback.usecase.interface";
+} from "./interfaces/oauth-signup-callback.usecase.interface";
 export { OAuthLoginCallbackUseCase } from "./oauth-login-callback.usecase";
 export { OAuthRequestUseCase } from "./oauth-request.usecase";
 export { OAuthSignupCallbackUseCase } from "./oauth-signup-callback.usecase";
