+++ apps/api/src/application/use-cases/oauth/interfaces/oauth-request.usecase.interface.ts
