+++ apps/api/src/application/use-cases/oauth/interfaces/oauth-signup-callback.usecase.interface.ts
