+++ apps/api/src/interface-adapter/gateway/oauth-provider/index.ts
@@ -1,6 +1,6 @@
 export type {
 	IOAuthProviderGateway,
 	AccountInfo,
-} from "./interface/oauth-provider.gateway.interface";
+} from "./interfaces/oauth-provider.gateway.interface";
 export { DiscordOAuthGateway } from "./providers/discord.gateway";
 export { OAuthProviderGateway } from "./oauth-provider.gateway";
