+++ apps/api/src/interface-adapter/gateway/oauth-provider/interfaces/oauth-provider.gateway.interface.ts
