+++ apps/api/src/interface-adapter/repositories/email-verification-code/index.ts
@@ -1,2 +1,2 @@
-export type { IEmailVerificationCodeRepository } from "./interface/email-verification-code.repository.interface";
+export type { IEmailVerificationCodeRepository } from "./interfaces/email-verification-code.repository.interface";
 export { EmailVerificationCodeRepository } from "./email-verification-code.repository";
