+++ apps/api/src/interface-adapter/repositories/oauth-account/index.ts
@@ -1,2 +1,2 @@
-export type { IOAuthAccountRepository } from "./interface/oauth-account.repository.interface";
+export type { IOAuthAccountRepository } from "./interfaces/oauth-account.repository.interface";
 export { OAuthAccountRepository } from "./oauth-account.repository";
