+++ apps/api/src/interface-adapter/repositories/oauth-account/interfaces/oauth-account.repository.interface.ts
