+++ apps/api/src/interface-adapter/repositories/oauth-account/oauth-account.repository.ts
@@ -1,7 +1,7 @@
 import { and, eq } from "drizzle-orm";
 import type { DrizzleService } from "../../../infrastructure/drizzle";
 import { OAuthAccount } from "../../../models/entities/oauth-account";
-import type { IOAuthAccountRepository } from "./interface/oauth-account.repository.interface";
+import type { IOAuthAccountRepository } from "./interfaces/oauth-account.repository.interface";
 
 export class OAuthAccountRepository implements IOAuthAccountRepository {
 	constructor(private readonly drizzleService: DrizzleService) {}
