+++ apps/api/src/interface-adapter/repositories/user-credential/interfaces/user-credential.repository.interface.ts
