+++ apps/api/src/interface-adapter/repositories/user/index.ts
@@ -1,2 +1,2 @@
 export { UserRepository } from "./user.repository";
-export type { IUserRepository } from "./interface/user.repository.interface";
+export type { IUserRepository } from "./interfaces/user.repository.interface";
