+++ apps/api/src/interface-adapter/repositories/user/interfaces/user.repository.interface.ts
