+++ apps/api/src/interface-adapter/repositories/user/user.repository.ts
@@ -2,7 +2,7 @@ import { eq } from "drizzle-orm";
 import type { DrizzleService } from "../../../infrastructure/drizzle";
 import type { Session } from "../../../models/entities/session";
 import { User } from "../../../models/entities/user";
-import type { IUserRepository } from "./interface/user.repository.interface";
+import type { IUserRepository } from "./interfaces/user.repository.interface";
 
 export class UserRepository implements IUserRepository {
 	constructor(private readonly drizzleService: DrizzleService) {}
